debian: Add genorig script
Currently, the orig tarball differs from the content of the upstream
branch, there are licensing issues and not the whole content if used for
the package (and most of untracked file are binaries). This adds a
genorig script that uses Files-Excluded from debian/copyright for
generating a new orig tarball. This tarball can later be used via gbp
import-orig.