z50-raspi3-firmware: Allow the kernel image and the initramfs to be configurable
This allows the kernel variable, from config.txt, to be configurable, so
we can start u-boot or barebox as second level bootloader and not only
the linux kernel. Moreover, in a case like this one, the usage of the
initramfs is not required, so add an option to disable its usage.