Add support for the armel architecture
Currently, this package supports armhf and arm64. The Rasberry PI 1 or
the Raspberry PI Zero are BCM2835-based boards. This SoC is based on an
armv6 ARM CPU, that is not compatible with the hard float ABI. On
upstream, this firmware is the binary blob to use for initializing the
hardware platform of all RPI (it does the job of a pre-bootloader or
SPL). For this reason, we add the armel architecture as supported
targets.